Moved genotype unit tests into new module.
